% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabf.R
\name{oddsTabf}
\alias{oddsTabf}
\title{Multiple Logistic Regression Models Summary Table (HTML)}
\usage{
oddsTabf(..., model_names = NULL)
}
\arguments{
\item{...}{Multiple fitted logistic regression model objects.}

\item{model_names}{An optional character vector of names for the models.}
}
\value{
An \code{htmlTable} object ready for viewing or knitting in an RMarkdown/Quarto document.
}
\description{
Similar to \code{oddsf()}, but immediately renders the output as a styled HTML table using the \code{htmlTable} package.
}
\examples{
# oddsTabf(mod1, mod2)
}
