% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabf.R
\name{oddsTabf}
\alias{oddsTabf}
\title{Two or more logistic regression model summary table with html}
\usage{
oddsTabf(...)
}
\arguments{
\item{...}{}
}
\description{
Two or more logistic regression model summary table with html
}
