% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabf.R
\name{tabf}
\alias{tabf}
\title{table 1 functon for public health (col sum)}
\usage{
tabf(dat1, stratas, catVars, conVars)
}
\arguments{
\item{conVars}{}
}
\description{
table 1 functon for public health (col sum)
}
