% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabf.R
\name{tabf}
\alias{tabf}
\title{Title}
\usage{
tabf(dat1, stratas, catVars, conVars)
}
\arguments{
\item{dat1}{a data set}

\item{stratas}{a variable}

\item{catVars}{a variable}

\item{conVars}{a variable}
}
\description{
Title
}
